+++ Sample/src/Sample.Api/Infrastructure/EnvironmentName.cs
@@ -1,9 +1,0 @@
-namespace Sample.Api.Infrastructure;
-
-public enum EnvironmentName
-{
-    Development,
-    AzureDev,
-    UAT,
-    Production
-}