+++ Sample/src/Sample.Api/ModuleConfig.cs
@@ -15,6 +15,12 @@ public static class ModuleConfig
         s.AddTransient<IReadRepository<SampleDbContext, WeatherForecast>, ReadRepository<SampleDbContext, WeatherForecast>>();
         s.AddTransient<ICudRepository<SampleDbContext, WeatherForecast>, CudRepository<SampleDbContext, WeatherForecast>>();
 
+        s.AddWithLogging<ICreateForecastHandler, CreateForecastHandler>();
+        s.AddWithLogging<IGetAllForecastsHandler, GetAllForecastsHandler>();
+        s.AddWithLogging<IGetForecastHandler, GetForecastHandler>();
+        s.AddWithLogging<IUpdateForecastHandler, UpdateForecastHandler>();
+        s.AddWithLogging<IDeleteForecastHandler, DeleteForecastHandler>();
+
         return builder;
     }