+++ src/CoreDesign.Identity/CoreDesign.Identity.Server/Identities/IdentityRecord.cs
@@ -9,6 +9,6 @@ public class IdentityRecord
     public string Name { get; set; } = string.Empty;
     public string GivenName { get; set; } = string.Empty;
     public string FamilyName { get; set; } = string.Empty;
-    public List<string> Roles { get; set; } = [];
+    public List<string> Permissions { get; set; } = [];
     public Dictionary<string, string> CustomClaims { get; set; } = [];
 }